[FIX] : Fetch 개선 KAN-999
